Rename service folder index.ts files to eponymous names
Both async-data-service and ui-service now expose their namespace via
the eponymous file (async-data-service.ts, ui-service.ts) per the
namespace rule. service/index.ts re-exports them, so consumers of
@adobe/data/service see no change.
